analyser/requirements.txt
python-dotenv
psycopg2
pandas
mkmsdk
